Add the project members calls to a new service
